Module bus_mapping::circuit_input_builder

This module contains the CircuitInputBuilder, which is an object that takes types from geth / web3 and outputs the circuit inputs.

Structs

Access

State Access caused by a transaction or an execution step

AccessSet

State and Code Access set.

Block

Circuit Input related to a block.

BlockContext

Context of a Block which can mutate in a Transaction.

BuilderClient

Struct that wraps a GethClient and contains methods to perform all the steps necessary to generate the circuit inputs for a block by querying geth for the necessary information and using the CircuitInputBuilder.

Call

Circuit Input related to an Ethereum Call

CallContext

Context of a Call.

CircuitInputBuilder

Builder to generate a complete circuit input from data gathered from a geth instance. This structure is the centre of the crate and is intended to be the only entry point to it. The CircuitInputBuilder works in several steps:

CircuitInputStateRef

Reference to the internal state of the CircuitInputBuilder in a particular ExecStep.

CopyEvent

Defines a copy event associated with EVM opcodes such as CALLDATACOPY, CODECOPY, CREATE, etc. More information: https://github.com/privacy-scaling-explorations/zkevm-specs/blob/master/specs/copy-proof.md.

CopyStep

Defines a single copy step in a copy event. This type is unified over the source/destination row in the copy table.

DynamicCParams

Unset Circuits Parameters, computed dynamically together with circuit witness generation.

ExecStep

An execution step of the EVM.

ExpEvent

Event representating an exponentiation a ^ b == d (mod 2^256).

ExpStep

Intermediary multiplication step, representing a * b == d (mod 2^256)

FixedCParams

Circuit Setup Parameters

Transaction

Result of the parsing of an Ethereum Transaction.

TransactionContext

Context of a Transaction which can mutate in an ExecStep.

Enums

AccessValue

State and Code Access with “keys/index” used in the access operation.

CallKind

Type of a CALL/CREATE* Function.

CodeSource

Source of the code in the EVM execution.

CopyDataType

Defines the various source/destination types for a copy event.

ExecState

Execution state

NumberOrHash

Defines an enum type that can hold either a number or a hash value.

Traits

CircuitsParams

Circuit Setup Parameters. These can be fixed/concrete or unset/dynamic.

Functions

build_state_code_db

Build a partial StateDB from step 3

get_call_memory_offset_length

Retrieve the memory offset and length of call.

get_create_init_code

Retrieve the init_code from memory for {CREATE, CREATE2}

get_state_accesses

Get State Accesses from TxExecTraces

keccak_inputs

Return all the keccak inputs used during the processing of the current block.

keccak_inputs_sign_verify

Generate the keccak inputs required by the SignVerify Chip from the signature datas.

keccak_inputs_tx_circuit

Generate the keccak inputs required by the Tx Circuit from the transactions.