Module bus_mapping::operation

Collection of structs and functions used to:

Define the internals of a MemoryOp, StackOp and StorageOp.

Define the actual operation types and a wrapper over them (the Operation enum).

Define structures that interact with operations such as OperationContainer.

Structs

AccountOp

Represents a change in the Account field implied by a BeginTx, EXTCODECOPY, EXTCODESIZE, BALANCE, SELFDESTRUCT, *CALL*, CREATE*, STOP, RETURN or REVERT step.

CallContextOp

Represents an CallContext read/write operation.

MemoryAddress

Represents a MemoryAddress of the EVM.

MemoryOp

Represents a READ/WRITE into the memory implied by an specific OpcodeId of the ExecStep.

Operation

Operation is a Wrapper over a type that implements Op with a RWCounter.

OperationContainer

The OperationContainer is meant to store all of the Operations that an ExecStep performs during its execution.

RWCounter

Wrapper type over usize which represents the global counter. The purpose of the RWCounter is to enforce that each Opcode/Instruction and Operation is unique and just executed once.

StackAddress

Represents a StackAddress of the EVM. The address range goes TOP -> DOWN (1024, 0].

StackOp

Represents a READ/WRITE into the stack implied by an specific OpcodeId of the ExecStep.

StartOp

Represent a Start padding operation

StorageOp

Represents a READ/WRITE into the storage implied by an specific OpcodeId of the ExecStep.

TargetIter

An iterator over the variants of Self

TxAccessListAccountOp

Represents a change in the Account AccessList implied by a BeginTx, EXTCODECOPY, EXTCODESIZE, EXTCODEHASH BALANCE, SELFDESTRUCT, *CALL* or CREATE* step.

TxAccessListAccountStorageOp

Represents a change in the Storage AccessList implied by an SSTORE or SLOAD step of the ExecStep.

TxLogOp

Represents TxLog read/write operation.

TxReceiptOp

Represents TxReceipt read/write operation.

TxRefundOp

Represents a change in the Transaction Refund AccessList implied by an SSTORE, STOP, RETURN or REVERT step of the ExecStep.

Enums

AccountField

Represents a field parameter of the Account that can be accessed via EVM execution.

CallContextField

Represents a field parameter of the CallContext that can be accessed via EVM execution.

OpEnum

Generic enum that wraps over all the operation types possible. In particular StackOp, MemoryOp and StorageOp.

RW

Marker that defines whether an Operation performs a READ or a WRITE.

Target

Enum used to differenciate between EVM Stack, Memory and Storage operations.

TxLogField

Represents a field parameter of the TxLog that can be accessed via EVM execution.

TxReceiptField

Represents a field parameter of the TxReceipt that can be accessed via EVM execution.

Traits

Op

Trait used for Operation Kinds.