Struct bus_mapping::operation::MemoryAddress

pub struct MemoryAddress(pub usize);

Represents a MemoryAddress of the EVM.

Tuple Fields
0: usize
Implementations

impl MemoryAddress

pub const fn zero() -> MemoryAddress

Returns the zero address for Memory targets.

pub fn to_le_bytes(self) -> [u8; 32]

Return the little-endian byte representation of the word as a 32-byte array.

pub fn to_be_bytes(self) -> [u8; 32]

Return the big-endian byte representation of the word as a 32-byte array.

pub fn from_le_bytes<T>(bytes: T) -> Result<MemoryAddress, Error>where T: AsRef<[u8]>,

Generate a MemoryAddress from the provided set of little-endian bytes.

pub fn from_be_bytes<T>(bytes: T) -> Result<MemoryAddress, Error>where T: AsRef<[u8]>,

Generate a MemoryAddress from the provided set of big-endian bytes.

pub fn map<F>(&self, f: F) -> MemoryAddresswhere F: FnOnce(usize) -> usize,

Apply a function to the contained value.